refactor: 모듈 추가, 패키지 및 파일 분리, 이동
- 컨텐츠 별 패키지 분리
