+++ features/carefullcontents/src/main/java/com/cases/carefull/features/carefullcontents/diagnosis/chatbot/ChatMessage.kt
