+++ features/carefullcontents/src/main/java/com/cases/carefull/features/carefullcontents/diagnosis/medicineInfo/MedicineInfoScreen.kt
