+++ features/carefullcontents/src/main/java/com/cases/carefull/features/carefullcontents/feed/Ranking.kt
@@ -1,4 +1,4 @@
-package com.cases.carefull.features.carefullmainui.screen.feed
+package com.cases.carefull.features.carefullcontents.feed
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Arrangement
