+++ features/carefullcontents/src/main/java/com/cases/carefull/features/carefullcontents/routine/Exercise.kt
@@ -1,4 +1,4 @@
-package com.cases.carefull.features.carefullmainui.screen.routine
+package com.cases.carefull.features.carefullcontents.routine
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -34,7 +34,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.Dialog
 import androidx.compose.ui.window.DialogProperties
-import com.cases.carefull.features.carefullmainui.theme.CarefullTheme
+import com.cases.carefull.features.carefullcommon.theme.CarefullTheme
 
 
 @Composable
