+++ features/carefullcontents/src/main/java/com/cases/carefull/features/carefullcontents/routine/SearchFood.kt
@@ -1,4 +1,4 @@
-package com.cases.carefull.features.carefullmainui.screen.routine
+package com.cases.carefull.features.carefullcontents.routine
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -35,7 +35,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.cases.carefull.features.carefullmainui.theme.CarefullTheme
+import com.cases.carefull.features.carefullcommon.theme.CarefullTheme
 
 @Composable
 fun SearchFood() {
