+++ features/carefullmainui/src/main/java/com/cases/carefull/features/carefullmainui/screen/Home.kt
@@ -17,7 +17,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.cases.carefull.features.carefullcommon.components.RowLine
-import com.cases.carefull.features.carefullmainui.theme.CarefullTheme
+import com.cases.carefull.features.carefullcommon.theme.CarefullTheme
 
 
 @Composable
